backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
pymongo==4.8.0
neo4j==5.23.0
confluent-kafka==2.5.0
pydantic-settings==2.4.0
httpx==0.27.0
sse-starlette==2.1.0
pandas==2.2.2
numpy==1.26.4
